requirements.txt
requests
python-dotenv
google-generativeai
mailjet-rest
supabase
matplotlib